backend/auth-service/requirements.txt
# Flask and extensions
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-Cors==4.0.0
Flask-WTF==1.2.1
Werkzeug==2.3.7

# Database
psycopg2-binary==2.9.7
SQLAlchemy==1.4.41

# Authentication
PyJWT==2.8.0
bcrypt==4.0.1
google-auth-oauthlib==1.0.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
redis==5.0.0
pydantic>=2.5.2
email-validator==2.0.0.post2
gunicorn==21.2.0
APScheduler==3.10.4

# Resilience
tenacity==8.2.3
circuitbreaker==1.4.0

# Monitoring and logging
prometheus-client==0.19.0
sentry-sdk==1.39.1
structlog==23.1.0
python-json-logger==2.0.7

# Service discovery and secrets
python-consul==1.1.0
kubernetes==28.1.0
hvac==1.1.1
boto3==1.29.6

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
responses==0.23.3

# Local shared package - installed via Dockerfile
# -e ../../meeting_shared 